+++ src/main/java/com/hello/spring/repository/HeroRepository.java
@@ -2,6 +2,8 @@ package com.hello.spring.repository;
 
 import com.hello.spring.model.Hero;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface HeroRepository extends JpaRepository<Hero,Integer> {
 }
