+++ src/main/java/com/hello/spring/repository/TeamRepository.java
@@ -2,6 +2,8 @@ package com.hello.spring.repository;
 
 import com.hello.spring.model.Team;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface TeamRepository extends JpaRepository<Team,Integer> {
 }
