+++ src/main/java/com/hello/spring/repository/UserRepository.java
@@ -2,6 +2,8 @@ package com.hello.spring.repository;
 
 import com.hello.spring.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface UserRepository extends JpaRepository<User,Integer> {
 }
